+++ sys/lib/libsa/exec.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: exec.c,v 1.22 2003/08/07 16:32:27 agc Exp $	*/
+/*	$NetBSD: exec.c,v 1.23 2003/08/31 22:40:48 fvdl Exp $	*/
 
 /*-
  * Copyright (c) 1982, 1986, 1990, 1993
@@ -88,7 +88,7 @@ exec(path, loadaddr, howto)
 		addr += sizeof(x);
 		x.a_text -= sizeof(x);
 	}
-	if (read(io, (char *)addr, x.a_text) != x.a_text)
+	if (read(io, (char *)addr, x.a_text) != (ssize_t)x.a_text)
 		goto shread;
 	addr += x.a_text;
 	if (N_GETMAGIC(x) == ZMAGIC || N_GETMAGIC(x) == NMAGIC)
@@ -97,13 +97,13 @@ exec(path, loadaddr, howto)
 
         /* Data */
 	printf("+%ld", x.a_data);
-	if (read(io, addr, x.a_data) != x.a_data)
+	if (read(io, addr, x.a_data) != (ssize_t)x.a_data)
 		goto shread;
 	addr += x.a_data;
 
         /* Bss */
 	printf("+%ld", x.a_bss);
-	for (i = 0; i < x.a_bss; i++)
+	for (i = 0; i < (int)x.a_bss; i++)
 		*addr++ = 0;
 
         /* Symbols */
@@ -112,7 +112,7 @@ exec(path, loadaddr, howto)
 	addr += sizeof(x.a_syms);
 	if (x.a_syms) {
 		printf("+[%ld", x.a_syms);
-		if (read(io, addr, x.a_syms) != x.a_syms)
+		if (read(io, addr, x.a_syms) != (ssize_t)x.a_syms)
 			goto shread;
 		addr += x.a_syms;
 	}
